% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmc_initialise.r
\name{cmc_initialise}
\alias{cmc_initialise}
\title{Set seed, loads packages and detach variables}
\usage{
cmc_initialise()
}
\value{
Nothing
}
\description{
Set seed, loads packages and detach variables
}
