% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmc_makeDraws.r
\name{cmc_makeDraws}
\alias{cmc_makeDraws}
\title{Creates a list of matrices containing draws. One matrix per dimension}
\usage{
cmc_makeDraws(cmc_inter_draws, cmc_intra_draws)
}
\arguments{
\item{cmc_inter_draws}{A list of arguments describing the inter-person
draws to create. The list must include the following elements:
- nDraws Scalar. Number of draws per individual per dimension.
- drawNames Vector of strings. Name of dimensions (i.e. random variables to draw for).
- normalDraws Vector of strings. Name of dimensions to turn into normal draws.
- drawType String. Can be 'halton', 'mlhs', or 'pmc'. Determines drawing method.}

\item{cmc_intra_draws}{A list of arguments describing the intra-person
draws to create. The list must include the following elements:
- nDraws Scalar. Number of draws per individual per dimension.
- drawNames Vector of strings. Name of dimensions (i.e. random variables to draw for).
- normalDraws Vector of strings. Name of dimensions to turn into normal draws.
- drawType String. Can be 'halton', 'mlhs', or 'pmc'. Determines drawing method.}
}
\value{
A list of arrays, one for each variable.
  Each array has dimensions nObs x nDrawsInter x nDrawsIntra
}
\description{
Creates a list of matrices containing draws. One matrix per dimension
}
