% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmc_mlhs.r
\name{cmc_mlhs}
\alias{cmc_mlhs}
\title{Generate random draws using the Modified Latin Hypercube algorithm.}
\usage{
cmc_mlhs(N, d, i)
}
\arguments{
\item{N}{The number of draws to generate in each dimension}

\item{d}{The number of dimensions to generate draws in}

\item{i}{The number of individuals to generate draws for}
}
\value{
A (N*i) x d matrix with random draws
}
\description{
Generate random draws using the Modified Latin Hypercube algorithm.
}
