% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmc_modeloutput.r
\name{cmc_modeloutput}
\alias{cmc_modeloutput}
\title{Prints to screen the output of a model previously estimated by runmodel()}
\usage{
cmc_modeloutput(model)
}
\arguments{
\item{model}{The model estimated using runmodel()}
}
\value{
A matrix of coefficients, s.d. and t-tests
}
\description{
Prints to screen the output of a model previously estimated by runmodel()
}
