% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmc_ologit.r
\name{cmc_ologit}
\alias{cmc_ologit}
\title{Calculates the probability of an ordered logit model}
\usage{
cmc_ologit(Y, X, lambda, tau)
}
\arguments{
\item{Y}{vector of dependent variables}

\item{X}{vector of a single explanatory variable (usually a latent variable)}

\item{lambda}{Scalar parameter. Multiplies X in the utility of the ordered logit}

\item{tau}{Vector of parameters. Thresholds.}
}
\value{
A vector of probabilities, one for each element of Y.
}
\description{
Calculates the probability of an ordered logit model
}
