% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmc_prepareCluster.r
\name{cmc_prepareCluster}
\alias{cmc_prepareCluster}
\title{Creates cluster and loads pieces of data in each cluster}
\usage{
cmc_prepareCluster()
}
\value{
Nothing (the cluster is loaded as a global variable)
}
\description{
Creates cluster and loads pieces of data in each cluster
}
