% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmc_prepareData.r
\name{cmc_prepareData}
\alias{cmc_prepareData}
\title{Orders data by respondent and prepares a data set to be used in model estimation}
\usage{
cmc_prepareData(idColName, choiceColName, alternatives, availColName = NA)
}
\arguments{
\item{idColName}{String indicating the name of the column containing respondents' ID.}

\item{choiceColName}{String indicating the name of the column containing respondents' choice (level coded).}

\item{alternatives}{Vector containing the codes (labels) used to register choices in the database, e.g.:c(1,2,3) or c('bus','train','plane'). If mdcev==1, then it must contain the column name of consumptions.}

\item{availColName}{String indicating the name of the columns containing respondents' availability. The order must be the same than in the choice coding. If ommited, full availability is assumed.}
}
\value{
Nothing. Data is loaded as a global variable.
}
\description{
Orders data by respondent and prepares a data set to be used in model estimation
}
