% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmc_runmodelGen.R
\name{cmc_runmodelGen}
\alias{cmc_runmodelGen}
\title{Estimate model using genetic algorithm rgenoud}
\usage{
cmc_runmodelGen(globalEnv)
}
\arguments{
\item{globalEnv}{A reference to the global environment (.GlobalEnv)}
}
\value{
Nothing. Model's results are loaded as a global variable.
}
\description{
Estimate model using genetic algorithm rgenoud
}
