% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmc_saveoutput.r
\name{cmc_saveoutput}
\alias{cmc_saveoutput}
\title{Estimates model}
\usage{
cmc_saveoutput(model)
}
\arguments{
\item{model}{Model estimated by runmodel() function}
}
\value{
Nothing. Model's results are loaded as a global variable.
}
\description{
Estimates model
}
