% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmc_searchInitialTheta.R
\name{cmc_searchInitialTheta}
\alias{cmc_searchInitialTheta}
\title{Search for better starting values by taking random pints around the user-defined starting point}
\usage{
cmc_searchInitialTheta(logLike)
}
\arguments{
\item{logLike}{log-likelihood function, must return a negative scalar}
}
\value{
The best theta found
}
\description{
Search for better starting values by taking random pints around the user-defined starting point
}
