% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmc_speedTest.r
\name{cmc_speedTest}
\alias{cmc_speedTest}
\title{Runs the cmc_probablities function several times on different number of cores and draws, and reports how long it takes.}
\usage{
cmc_speedTest(nDrawsTry = c(10, 50, 100), nCoresTry = NA, nRep = 30)
}
\arguments{
\item{nDrawsTry}{Vector of number of draws to try (ignored if cmc_control$mixing==0)}

\item{nCoresTry}{Vector of number of cores to try (defaults to 1:maximum_number_of_cores_detected)}

\item{nRep}{Scalar. Number of times cmc_probabilities is calculated (defaults to 30)}
}
\value{
A matrix indicating the time taken to estimate nRep times the cmc_probabilities function.
}
\description{
Runs the cmc_probablities function several times on different number of cores and draws, and reports how long it takes.
}
