% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmc_validatecontrol.r
\name{cmc_validatecontrol}
\alias{cmc_validatecontrol}
\title{Validates the cmc_control list of estimation parameters. Sets default values for the ommited ones.}
\usage{
cmc_validatecontrol()
}
\value{
Nothing. The validated cmc_control list is stored as a global variable.
}
\description{
Validates the cmc_control list of estimation parameters. Sets default values for the ommited ones.
}
