% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltamethod.r
\name{deltamethod}
\alias{deltamethod}
\title{Applies the delta method to calculate the standard error of transformations of parameters}
\usage{
deltamethod(par1, par2)
}
\arguments{
\item{par1}{String. Name of the first parameter.}

\item{par2}{String. Name of the second parameter.}
}
\value{
Matrix indicating the value and s.e. of the addition, substraction and ratio of the two parameters.
}
\description{
Applies the delta method to calculate the standard error of transformations of parameters
}
