% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNIter.r
\name{readNIter}
\alias{readNIter}
\title{Reads the number of iterations from the file written with the parameters' values at each iteration}
\usage{
readNIter()
}
\value{
Number of iterations written in the file.
}
\description{
Reads the number of iterations from the file written with the parameters' values at each iteration
}
