% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeBeta.r
\name{writeBeta}
\alias{writeBeta}
\title{Writes the vector [beta,ll] to a file called modelname_iterations.csv}
\usage{
writeBeta(beta, ll)
}
\arguments{
\item{beta}{vector of parameters to be written}

\item{ll}{scalar representing the loglikelihood of the whole model}
}
\value{
Nothing.
}
\description{
Writes the vector [beta,ll] to a file called modelname_iterations.csv
}
