% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeF12.r
\name{writeF12}
\alias{writeF12}
\title{Writes an F12 file with the results of a model estimation.}
\usage{
writeF12()
}
\value{
Nothing.
}
\description{
Writes an F12 file with the results of a model estimation.
}
